Update command-line args parser
