+++ internal/exec/describe.go
@@ -23,7 +23,10 @@ func ExecuteDescribeComponent(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	err = c.InitConfig(stack)
+	var configAndStacksInfo c.ConfigAndStacksInfo
+	configAndStacksInfo.Stack = stack
+
+	err = c.InitConfig(configAndStacksInfo)
 	if err != nil {
 		return err
 	}
